增加电机模块的初版实现代码，增加PID算法的接入，已能够初步执行相关控制过程

+++ device/motor/include/motor_encoder.h
@@ -10,8 +10,7 @@ extern "C" {
 
 struct MOTOR_ENCODER;
 
-
-int32_t motor_encoder_update(struct MOTOR_ENCODER *encoder);
+float motor_encoder_update(struct MOTOR_ENCODER *encoder);
 
 #ifdef __cplusplus
 }
